requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.10
black==25.1.0
cffi==1.17.1
cfgv==3.4.0
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
dill==0.4.0
distlib==0.3.9
Django==5.2.1
django-ninja==1.4.1
filelock==3.18.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
identify==2.6.10
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
structlog==25.3.0
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1
